// src/app/components/lit-dialect-view/interfaces/lit-dialect-view.interface.ts
export interface ILitDialectView {
      dialectId: string | null;    // System.String
      dialectName: string | null;    // System.String
      iso3CntrRef: string | null;    // System.String
      iso2CntrRef: string | null;    // System.String
      iso3LngRef: string | null;    // System.String
      iso2LngRef: string | null;    // System.String
      cCountryName: string | null;    // System.String
      lLanguageName: string | null;    // System.String
}